separated project to multiple packages and added more documentation
